+++ tests/Framework/TestResources.php
@@ -836,6 +836,18 @@ class TestResources
         return $sample;
     }
 
+    public static function createMessageSample()
+    {
+        $sample = array();
+        $sample['QueueMessage']['MessageId']       = '5974b586-0df3-4e2d-ad0c-18e3892bfca2';
+        $sample['QueueMessage']['InsertionTime']   = 'Fri, 09 Oct 2009 21:04:30 GMT';
+        $sample['QueueMessage']['ExpirationTime']  = 'Fri, 16 Oct 2009 21:04:30 GMT';
+        $sample['QueueMessage']['PopReceipt']      = 'YzQ4Yzg1MDItYTc0Ny00OWNjLTkxYTUtZGM0MDFiZDAwYzEw';
+        $sample['QueueMessage']['TimeNextVisible'] = 'Fri, 09 Oct 2009 23:29:20 GMT';
+
+        return $sample;
+    }
+
     public static function listMessagesMultipleMessagesSample()
     {
         $sample = array();